+++ src/firebase.js
@@ -1,31 +1,19 @@
-// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getAnalytics } from "firebase/analytics";
-import { getFirestore } from 'firebase/firestore';
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
+import { getFirestore } from "firebase/firestore";
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
 const firebaseConfig = {
-  apiKey: "AIzaSyC61Knl5cLtztAk13LH8R2Sfp3QAeNyG5I",
-  authDomain: "chsdhkd-c2072.firebaseapp.com",
-  projectId: "chsdhkd-c2072",
-  storageBucket: "chsdhkd-c2072.firebasestorage.app",
-  messagingSenderId: "960905359501",
-  appId: "1:960905359501:web:3018ba9884b6dbb36472c4",
-  measurementId: "G-TTBG49NVWL"
+    apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+    authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+    projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
+    storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
+    messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
+    appId: process.env.REACT_APP_FIREBASE_APP_ID,
+    measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
 };
 
-// Initialize Firebase
 const app = initializeApp(firebaseConfig);
 const analytics = getAnalytics(app);
 
-
-
-
-// 🧠 Firestore 객체 생성
 const db = getFirestore(app);
-
-// ✅ db 내보내기
 export { db };
