--- src/pages/Home.js
@@ -1,17 +1,0 @@
-import React, { useEffect, useState } from 'react';
-
-import './Home.css';
-
-const Home= () => {
-  
-
-  return (
-    <div >
-      <h1>홈 화면</h1>
-      <img src="./images/home.webp" alt="home" width={830} height={700} />
-      
-    </div>
-  );
-};
-
-export default Home;
