+++ src/pages/Home/index.jsx
@@ -1,0 +1,14 @@
+import React, { useEffect, useState } from "react";
+
+import "./index.css";
+
+const Home = () => {
+    return (
+        <div>
+            <h1>홈 화면</h1>
+            <img src="./images/home.webp" alt="home" width={830} height={700} />
+        </div>
+    );
+};
+
+export default Home;
